feat: Add is_boardgame column to Listings table
We need to store which listings don't represent boardgames (ex: dices,
binders, decks of cards). This will serve to reduce the unnecesary
identification calls on listings that aren't boardgames and filter them
out in the dashboard index views.

The option to make this column an listing_type or product_type Enum was
considered and it might prove more flexible, but we're not expecting any
near-future usecase and the migration from a boolean to an Enum should