feat(SearchMethod::SearchResult): add data class to represent search results
We were using Bgg::SearchResult to represent search results, but that class
belonged to the BGG API module, which should be isolated from the rest of the
application. Therefore, we created a new class SearchMethod::SearchResult to
represent search results.

This class includes new attributes:
- rank (Integer): the rank of the boardgame in the database
- similarity (Float): the similarity of the search result to the query

The idea is to later use these attributes to sort each search result
with a custom sorting algorithm.